Move Enemies to separate class and organize sprites
