+++ Sprites/barrelBeige.png
