+++ Sprites/barrelRed.png
