+++ Sprites/coin_07.png
