+++ Sprites/explosion3.png
