+++ Sprites/letter_C.png
