+++ Sprites/shotThin.png
