+++ Sprites/specialBarrel6.png
