+++ Sprites/tankBody_huge.png
