+++ Sprites/tank_blue.png
