+++ Sprites/tank_darkLarge.png
