+++ Sprites/tank_huge.png
