+++ Sprites/tank_red.png
