+++ Sprites/tank_sand.png
